change size variable to count
Because we want to use the copy functions from the base class ArrayList but the size there is the number of elements in the arraylist, in the heap the size variable was keeping track of the last filled element. Off by one. Change variable name to count to be more ambiguous.

Override Constructor because count needs to start at 1.
Override isEmpty since count starts one count  <= 1 means empty data structure.
Passes tests